+++ database/migrations/2019_12_08_093650_create_user_account_table.php
@@ -16,9 +16,9 @@ class CreateUserAccountTable extends Migration
         Schema::create('user_account', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('user_id')->unique();
-            $table->unsignedInteger('balance')->default(0)->comment('余额');
-            $table->unsignedInteger('withdrawn')->default(0)->comment('已提现');
-            $table->unsignedInteger('cash_in')->default(0)->comment('提现中');
+            $table->double('balance',10,2)->default(0)->comment('余额');
+            $table->double('withdrawn',10,2)->default(0)->comment('已提现');
+            $table->double('cash_in',10,2)->default(0)->comment('提现中');
             $table->timestamps();
         });
     }
